refactor(pick/onset): Rename onset/pick subpackages for clarity
Rename the onset and pick subpakckages to improve clarity in preparation for a
new routine for independently picking phase arrivals after locate has been
run.

onset -> onsets (motivation: this subpackage will expand to host multiple
types of onset functions)
pick -> pickers (motivation: this subpackage will expand to host multiple types
of pickers)

Will cause some changes to imports, but these are going to be affected by the
package name change anyway.
