+++ quakemigrate/signal/onsets/onset.py
