+++ quakemigrate/signal/onsets/staltaonset.py
